AI/requirements.txt
# Editable install with no version control (ai_video_analysis==0.1)
# -e d:\others\pregnancy-tracker
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.3
fpdf==1.7.2
fsspec==2024.12.0
gradio==5.12.0
gradio_client==1.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
opencv-python==4.10.0.84
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==6.1.1
py-cpuinfo==9.0.0
pybase64==1.4.0
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruff==0.9.1
safehttpx==0.1.6
scipy==1.15.1
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
supervision==0.25.1
sympy==1.13.1
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.59
ultralytics-thop==2.0.13
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.1
Werkzeug==3.1.3
